with_monitor_returns_true
