+++ lumen_runtime/src/otp/erlang/demonitor_2/test/with_reference/with_flush_and_info_options/with_monitor.rs
@@ -2,40 +2,7 @@ use super::*;
 
 #[test]
 fn returns_true() {
-    with_process_arc(|monitoring_arc_process| {
-        let monitored_arc_process = process::test(&monitoring_arc_process);
-
-        let monitor_reference = monitor_2::native(
-            &monitoring_arc_process,
-            r#type(),
-            monitored_arc_process.pid_term(),
-        )
-        .unwrap();
-
-        let monitored_monitor_count_before = monitor_count(&monitored_arc_process);
-        let monitoring_monitored_count_before = monitored_count(&monitoring_arc_process);
-
-        assert_eq!(
-            native(
-                &monitoring_arc_process,
-                monitor_reference,
-                options(&monitoring_arc_process)
-            ),
-            Ok(true.into())
-        );
-
-        let monitored_monitor_count_after = monitor_count(&monitored_arc_process);
-        let monitoring_monitored_count_after = monitored_count(&monitoring_arc_process);
-
-        assert_eq!(
-            monitored_monitor_count_after,
-            monitored_monitor_count_before - 1
-        );
-        assert_eq!(
-            monitoring_monitored_count_after,
-            monitoring_monitored_count_before - 1
-        );
-    });
+    with_monitor_returns_true(options);
 }
 
 #[test]
